fixed several warnings from Xcode 4 LLVM compiler; fixed bug #920

+++ 3rdparty/libjasper/jas_debug.c
@@ -125,7 +125,7 @@ int jas_memdump(FILE *out, void *data, size_t len)
 	uchar *dp;
 	dp = data;
 	for (i = 0; i < len; i += 16) {
-		fprintf(out, "%04x:", i);
+		fprintf(out, "%04x:", (int)i);
 		for (j = 0; j < 16; ++j) {
 			if (i + j < len) {
 				fprintf(out, " %02x", dp[i + j]);
